Person/Group/Location crosslinking with Occurrences
Occurrences can now have people, groups, and locations linked to them
the same way sources can and they will display on the occurrence entry
in a similar fashion.

If a person, group, or location is linked to an occurrence, that
occurrence will now show on the person, group, or location page for
informational purposes.